change the whole project into a better structure
