--- server/src/resolvers/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "resolvers",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "jQuery": "^1.7.4",
-    "jsdom": "^14.0.0",
-    "json_merger": "^1.2.0",
-    "mockjs": "^1.0.1-beta3"
-  }
-}
